fix: Fixes bug with babel.dates.format_datetime
Bable defaults to looking for an environment variable called
LC_TIME. Because cron must've not set this on my mac, it was
causing Locale.parse(locale) to return None becuase locale is
set to the LC_TIME env var. This fix sets
tzinfo=get_timezone('US/Eastern'), locale='en_US' in the
format_datetime function in order to fix this bug.