refactor: Adds _edit_or_create_venue_or_artist function
This commit adds _edit_or_create_venue_or_artist which is now
used in all other functions which, you guessed it, edit or
create an artist or venue. This function accounts for the
relation between venues and genres, and artists and genres.
This means that when genres are added to a venue or artist,
the proper database relationship in the ArtistGenre and
VenueGenre table is added.